requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Vector databases (choose one)
pinecone-client==3.0.2
qdrant-client==1.7.3
weaviate-client==4.4.1
pymilvus==2.3.5

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0  # for PostgreSQL
aiosqlite==0.19.0  # for SQLite

# Redis
redis==5.0.1
aioredis==2.0.1

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Embeddings (local model)
sentence-transformers==2.3.1
transformers==4.37.2
torch==2.1.2

# LLM
langchain==0.1.6
langchain-community==0.0.19

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3